backend/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
